bfloat16

+++ scripts/image/infer.py
@@ -35,7 +35,7 @@ url = "COCO_train2014_000000106644.jpg"
 image = Image.open(url).convert("RGB")
 print("image processor: ", image_processor)
 image_tensor = process_images([image], image_processor, model.config)
-image_tensor = [_image.to(dtype=torch.float16, device=device) for _image in image_tensor]
+image_tensor = [_image.to(dtype=torch.bfloat16, device=device) for _image in image_tensor]
 print("Image tensor: ", image_tensor[0].shape)
 conv_template = "qwen_1_5"  # Make sure you use correct chat template for different models
 question = DEFAULT_IMAGE_TOKEN + "\nWhat are the differences between a muffin and a cupcake?"
